requirements.txt
# Core Dependencies
numpy>=1.21.0
sentence-transformers>=2.2.0
pymupdf>=1.23.0  # fitz for PDF handling
python-docx # for word file handling

# LLM Integration
openai>=1.0.0
python-dotenv>=1.0.0

# Improved Features
tenacity>=8.2.0  # Retry logic
tqdm>=4.65.0     # Progress bars

# Data Handling
pandas>=1.5.0

# Deep Learning Framework (for Sentence Transformers)
torch>=2.0.1

